add more oop to duke
